+++ client/web/src/search/results/streaming/sidebar/SearchSidebar.module.scss
@@ -4,27 +4,3 @@
     width: 15.5rem;
     flex-shrink: 0;
 }
-
-.sidebar-section {
-    &__list {
-        list-style-type: none;
-        padding: 0;
-        margin-top: 0.5rem;
-    }
-
-    &__list-item {
-        display: block;
-        font-size: 0.75rem;
-        padding: 0.25rem 0.5rem;
-        border-radius: 3px;
-
-        &:hover {
-            background-color: var(--color-bg-2);
-            text-decoration: none;
-        }
-
-        &:active {
-            background-color: var(--color-bg-3);
-        }
-    }
-}
